Correct location of tasks
